requirements.txt
# qdrant-client
# langchain
# openai
# requests
# python-dotenv
# langchain-community
# gradio


qdrant-client>=1.12.0
langchain==0.0.200
langchain-community==0.0.1
openai==0.27.8
requests==2.31.0
python-dotenv==1.0.0
gradio==3.35.2